10.5号更新

+++ app/src/main/java/com/example/classchat/Activity/Activity_TodoDetail.java
@@ -86,9 +86,6 @@ public class Activity_TodoDetail extends AppCompatActivity {
     //周数数组
     List<Integer> weeksnum = new ArrayList<>();
 
-    //记录编辑状态flag,0未编辑，1编辑
-    private int editting = 0;
-
     //删除模式选择
     private AlertDialog builder = null;
     private boolean isSeeAll = false;
@@ -126,7 +123,6 @@ public class Activity_TodoDetail extends AppCompatActivity {
                 .path(getCacheDir(this))
                 .getCache("BeginClassTime",String.class);
 
-        Log.e("TAG", mBeginClassTime);
         if(mBeginClassTime == null || mBeginClassTime.length() <= 0){
             mBeginClassTime = getDate(0) + " 00:00:00";
         }
@@ -191,7 +187,7 @@ public class Activity_TodoDetail extends AppCompatActivity {
         content.setEnabled(false);
         back = findViewById(R.id.memo_edit_back);
         delete = findViewById(R.id.memo_delete);
-        save = findViewById(R.id.add_todo_button);
+        save = findViewById(R.id.memo_done);
         save.setVisibility(View.GONE);
         edit = findViewById(R.id.memo_edit);
         edit.setVisibility(View.VISIBLE);
@@ -238,11 +234,15 @@ public class Activity_TodoDetail extends AppCompatActivity {
                                                                 case 1:
                                                                     //todo
                                                                     address = "http://106.12.105.160:8081/deletesametodoitem";
-                                                                    deleteAlarm(memo);
+                                                                    if (memo.isClock() == true){
+                                                                        deleteAlarm(memo);
+                                                                    }
                                                                     break;
                                                                 default:
                                                                     address = "http://106.12.105.160:8081/deletetodoitem";
-                                                                    deleteAlarm_one(week, memo);
+                                                                    if (memo.isClock() == true){
+                                                                        deleteAlarm_one(week, memo);
+                                                                    }
                                                                     break;
                                                             }
 
@@ -316,123 +316,112 @@ public class Activity_TodoDetail extends AppCompatActivity {
         edit.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
-                if(editting == 0){
-                    editting ++;
-                    save.setVisibility(View.VISIBLE);
-                    content.setEnabled(true);
-                    isClock.setEnabled(true);
-                    title.setEnabled(true);
-                    setWeek.setEnabled(true);
-                    setTime.setEnabled(true);
-                    setTimeSlot.setEnabled(true);
-                    isClock.setOnCheckedChangeListener(new CompoundButton.OnCheckedChangeListener() {
-                        @Override
-                        public void onCheckedChanged(CompoundButton buttonView, boolean isChecked) {
-                            bisClock = isChecked;
-                        }
-                    });
-                    //周数多选框
-                    mutilChoicebuilder = new AlertDialog.Builder(Activity_TodoDetail.this);
-                    mutilChoicebuilder.setTitle("选择周数");
-                    mutilChoicebuilder.setMultiChoiceItems(weeks, weeksChecked, new DialogInterface.OnMultiChoiceClickListener() {
-                        @Override
-                        public void onClick(DialogInterface dialog, int which, boolean isChecked) { }
-                    });
-                    mutilChoicebuilder.setPositiveButton("确定", new DialogInterface.OnClickListener() {
-                        @Override
-                        public void onClick(DialogInterface dialog, int which) {
 
-                            if (weeksnum.size() > 0){
-                                setWeek.setText(setWeekTextView());
-                            }else{
-                                //没有选择
-                                Toast.makeText(Activity_TodoDetail.this, "未选择周数!", Toast.LENGTH_SHORT).show();
+                edit.setVisibility(View.GONE);
+                save.setVisibility(View.VISIBLE);
+                content.setEnabled(true);
+                isClock.setEnabled(true);
+                title.setEnabled(true);
+                setWeek.setEnabled(true);
+                setTime.setEnabled(true);
+                setTimeSlot.setEnabled(true);
+                isClock.setOnCheckedChangeListener(new CompoundButton.OnCheckedChangeListener() {
+                    @Override
+                    public void onCheckedChanged(CompoundButton buttonView, boolean isChecked) {
+                        bisClock = isChecked;
+                    }
+                });
+                //周数多选框
+                mutilChoicebuilder = new AlertDialog.Builder(Activity_TodoDetail.this);
+                mutilChoicebuilder.setTitle("选择周数");
+                mutilChoicebuilder.setMultiChoiceItems(weeks, weeksChecked, new DialogInterface.OnMultiChoiceClickListener() {
+                    @Override
+                    public void onClick(DialogInterface dialog, int which, boolean isChecked) { }
+                });
+                mutilChoicebuilder.setPositiveButton("确定", new DialogInterface.OnClickListener() {
+                    @Override
+                    public void onClick(DialogInterface dialog, int which) {
+
+                        if (weeksnum.size() > 0){
+                            setWeek.setText(setWeekTextView());
+                        }else{
+                            //没有选择
+                            Toast.makeText(Activity_TodoDetail.this, "未选择周数!", Toast.LENGTH_SHORT).show();
+                        }
+
+                    }
+                });
+
+                mutilChoicebuilder.setNegativeButton("取消", new DialogInterface.OnClickListener() {
+                    @Override
+                    public void onClick(DialogInterface dialog, int which) {
+
+                    }
+                });
+
+                setWeek.setOnClickListener(new View.OnClickListener() {
+                    @Override
+                    public void onClick(View v) {
+                        mutilChoicebuilder.show();
+                    }
+                });
+
+                setTime.setOnClickListener(new View.OnClickListener() {
+                    @RequiresApi(api = Build.VERSION_CODES.M)
+                    @Override
+                    public void onClick(View v) {
+                        show_timePicker();
+                    }
+                });
+
+                setTimeSlot.setOnClickListener(new View.OnClickListener() {
+                    @Override
+                    public void onClick(View v) {
+                        showTimeSlotPicker();
+                    }
+                });
+
+                save.setOnClickListener(new View.OnClickListener() {
+                    @Override
+                    public void onClick(View v) {
+                        String newWeekString = setWeekString(weeksnum);
+                        String oldWeekString = setWeekString(memo.getWeekChosen());
+
+                        final RequestBody requestBody = new FormBody.Builder()
+                                .add("userID", userID)
+                                .add("todoTitle", title.getText().toString())
+                                .add("newWeekList", newWeekString)
+                                .add("oldWeekList", oldWeekString)
+                                .add("dayChosen", dayOfweek + "")
+                                .add("timeSlot", timeslot + "")
+                                .add("detailTime", hour + " " + minute_)
+                                .add("isClock", bisClock + "")
+                                .add("content", content.getText().toString())
+                                .add("todoItemID", memo.getTodoItemID())
+                                .build();   //构建请求体
+//                        //TODO
+                        Util_NetUtil.sendOKHTTPRequest("http://106.12.105.160:8081/updatetodoitem", requestBody, new okhttp3.Callback() {
+                            @Override
+                            public void onResponse(Call call, Response response) throws IOException {
+                                // 得到服务器返回的具体内容
+                                boolean responseData = Boolean.parseBoolean(response.body().string());
+                                Message message = new Message();
+                                if (responseData) {
+                                    message.what = SAVE_SUCCESS;
+                                    handler.sendMessage(message);
+                                } else {
+                                    message.what = SAVE_FAILED;
+                                    handler.sendMessage(message);
+                                }
                             }
 
-                        }
-                    });
-
-                    mutilChoicebuilder.setNegativeButton("取消", new DialogInterface.OnClickListener() {
-                        @Override
-                        public void onClick(DialogInterface dialog, int which) {
-
-                        }
-                    });
-
-                    setWeek.setOnClickListener(new View.OnClickListener() {
-                        @Override
-                        public void onClick(View v) {
-                            mutilChoicebuilder.show();
-                        }
-                    });
-
-                    setTime.setOnClickListener(new View.OnClickListener() {
-                        @RequiresApi(api = Build.VERSION_CODES.M)
-                        @Override
-                        public void onClick(View v) {
-                            show_timePicker();
-                        }
-                    });
-
-                    setTimeSlot.setOnClickListener(new View.OnClickListener() {
-                        @Override
-                        public void onClick(View v) {
-                            showTimeSlotPicker();
-                        }
-                    });
-
-                    save.setOnClickListener(new View.OnClickListener() {
-                        @Override
-                        public void onClick(View v) {
-                            String newWeekString = setWeekString(weeksnum);
-                            String oldWeekString = setWeekString(memo.getWeekChosen());
-
-                            final RequestBody requestBody = new FormBody.Builder()
-                                    .add("userID", userID)
-                                    .add("todoTitle", title.getText().toString())
-                                    .add("newWeekList", newWeekString)
-                                    .add("oldWeekList", oldWeekString)
-                                    .add("dayChosen", dayOfweek + "")
-                                    .add("timeSlot", timeslot + "")
-                                    .add("detailTime", hour + " " + minute_)
-                                    .add("isClock", bisClock + "")
-                                    .add("content", content.getText().toString())
-                                    .add("todoItemID", memo.getTodoItemID())
-                                    .build();   //构建请求体
-//                        //TODO
-                            Util_NetUtil.sendOKHTTPRequest("http://106.12.105.160:8081/updatetodoitem", requestBody, new okhttp3.Callback() {
-                                @Override
-                                public void onResponse(Call call, Response response) throws IOException {
-                                    // 得到服务器返回的具体内容
-                                    boolean responseData = Boolean.parseBoolean(response.body().string());
-                                    Message message = new Message();
-                                    if (responseData) {
-                                        message.what = SAVE_SUCCESS;
-                                        handler.sendMessage(message);
-                                    } else {
-                                        message.what = SAVE_FAILED;
-                                        handler.sendMessage(message);
-                                    }
-                                }
-
-                                @Override
-                                public void onFailure(@NonNull Call call, @NonNull IOException e) {
-                                    // 在这里对异常情况进行处理
-                                }
-                            });
-                        }
-                    });
-                }
-                else {
-                    editting --;
-                    save.setVisibility(View.GONE);
-                    content.setEnabled(false);
-                    isClock.setEnabled(false);
-                    title.setEnabled(false);
-                    setWeek.setEnabled(false);
-                    setTime.setEnabled(false);
-                    setTimeSlot.setEnabled(false);
-                }
+                            @Override
+                            public void onFailure(@NonNull Call call, @NonNull IOException e) {
+                                // 在这里对异常情况进行处理
+                            }
+                        });
+                    }
+                });
             }
         });
     }
@@ -626,13 +615,20 @@ public class Activity_TodoDetail extends AppCompatActivity {
         public void handleMessage(Message msg) {
             switch (msg.what) {
                 case SAVE_SUCCESS:
+                    save.setVisibility(View.GONE);
+                    content.setEnabled(false);
+                    isClock.setEnabled(false);
+                    title.setEnabled(false);
+                    setWeek.setEnabled(false);
+                    setTime.setEnabled(false);
+                    setTimeSlot.setEnabled(false);
+                    edit.setVisibility(View.VISIBLE);
                     Util_ToastUtils.showToast(Activity_TodoDetail.this, "修改成功！");
                     deleteAlarm(memo);
                     if ( bisClock == true ){
                         Log.e("setAlarm", "start");
                         setAlarm();
                     }
-                    finish();
                     break;
                 case SAVE_FAILED:
                 case DELETE_FAILED:
@@ -794,14 +790,16 @@ public class Activity_TodoDetail extends AppCompatActivity {
      */
     private void deleteAlarm_one(int week, Object_TodoList memo){
         Log.e("deleteAlarm_one", "start");
-        for(int i = 0; i < alarm.size() ; i++){
-            Object_Todo_Broadcast_container obj = JSON.parseObject(alarm.get(i), Object_Todo_Broadcast_container.class);
-            String detail_time = obj.getHour() + " " + obj.getMinute() ;
-            Log.e("detail_time", detail_time);
-            if (obj.getTitle().equals(memo.getTodoTitle()) && obj.getDetail().equals(memo.getContent()) && detail_time.equals(memo.getDetailTime()) && week == obj.getWeek()){
-                AlarmTimer alarmTimer = new AlarmTimer(obj);
-                alarmTimer.cancelAlarmTimer(this);
-                Log.e("delete_id", obj.getId() + "");
+        if(alarm != null){
+            for(int i = 0; i < alarm.size() ; i++){
+                Object_Todo_Broadcast_container obj = JSON.parseObject(alarm.get(i), Object_Todo_Broadcast_container.class);
+                String detail_time = obj.getHour() + " " + obj.getMinute() ;
+                Log.e("detail_time", detail_time);
+                if (obj.getTitle().equals(memo.getTodoTitle()) && obj.getDetail().equals(memo.getContent()) && detail_time.equals(memo.getDetailTime()) && week == obj.getWeek()){
+                    AlarmTimer alarmTimer = new AlarmTimer(obj);
+                    alarmTimer.cancelAlarmTimer(this);
+                    Log.e("delete_id", obj.getId() + "");
+                }
             }
         }
     }
@@ -811,14 +809,16 @@ public class Activity_TodoDetail extends AppCompatActivity {
      */
     private void deleteAlarm(Object_TodoList memo){
         Log.e("deleteAlarm", "start");
-        for(int i = 0; i < alarm.size() ; i++){
-            Object_Todo_Broadcast_container obj = JSON.parseObject(alarm.get(i), Object_Todo_Broadcast_container.class);
-            String detail_time = obj.getHour() + " " + obj.getMinute() ;
-            Log.e("detail_time", detail_time);
-            if (obj.getTitle().equals(memo.getTodoTitle()) && obj.getDetail().equals(memo.getContent()) && detail_time.equals(memo.getDetailTime())){
-                AlarmTimer alarmTimer = new AlarmTimer(obj);
-                alarmTimer.cancelAlarmTimer(this);
-                Log.e("delete_id", obj.getId() + "");
+        if (alarm != null){
+            for(int i = 0; i < alarm.size() ; i++){
+                Object_Todo_Broadcast_container obj = JSON.parseObject(alarm.get(i), Object_Todo_Broadcast_container.class);
+                String detail_time = obj.getHour() + " " + obj.getMinute() ;
+                Log.e("detail_time", detail_time);
+                if (obj.getTitle().equals(memo.getTodoTitle()) && obj.getDetail().equals(memo.getContent()) && detail_time.equals(memo.getDetailTime())){
+                    AlarmTimer alarmTimer = new AlarmTimer(obj);
+                    alarmTimer.cancelAlarmTimer(this);
+                    Log.e("delete_id", obj.getId() + "");
+                }
             }
         }
     }
